[xml-writter] Speedup function_type::get_cached_name
It looks like due to a typo, we are never caching the name of the
function_type, so we are computing it all the time, *OOOPS*.  So this
is having an impact when comparing instance of function_type during
de-duplication at abixml writting time.

Things are faster now, thanks to this patch.

	* src/abg-ir.cc (function_type::get_cached_name): Really cache the
	computed name of function_type instances.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>