Remove the elf_symbol::get_value property
Now that there are proper facilities to lookup ELF symbols inside the
ELF/DWARF reader and get a native GElf_Sym type instance (from
libelf), we don't need to carry the value of the symbol (that is
relevant only that low level anyway) in the abigail::ir::elf_symbol
type.

This patch removes that property.

	* include/abg-ir.h (elf_symbol::{elf_symbol, create}): Remove the
	'val' parameter.
	* src/abg-dwarf-reader.cc (elf_symbol::get_value): Remove this
	member function declaration.
	(lookup_symbol_from_sysv_hash_tab)
	(lookup_symbol_from_gnu_hash_tab, lookup_symbol_from_symtab)
	(create_default_var_sym, create_default_fn_sym)
	(read_context::lookup_elf_symbol_from_index): Adjust calls to
	creating elf_symbol instances.
	* src/abg-ir.cc (elf_symbol::priv::value_): Remove this data
	member.
	(elf_symbol::{priv::priv, elf_symbol, create): Adjust.
	* src/abg-reader.cc (build_elf_symbol): Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>