dwarf-reader: read_corpus_from_elf: unconditionally load elf properties
When loading the corpus from elf while specifying a whitelist, we might
be able to ignore the symbol table. In any case we have to load the elf
properties into the context, such as the binary's architecture.
Otherwise they are missing from the internal / xml representation.
Previously, elf properties were not loaded when a whitelist was
specified. Fix that.

	* src/abg-dwarf-reader.cc (read_corpus_from_elf):
	unconditionally load elf properties into context

Signed-off-by: Matthias Maennich <maennich@google.com>