tools-utils: Drop redefinition of fopen when BAD_FTS is defined
When the BAD_FTS macro was defined at compile time (to handle the use
of fts.h on glibc < 2.23), we needed to re-map fopen to fopen64 if
_FILE_OFFSET_BITS=64 was defined.  We do not need this anymore because
we don't use fopen in that module anymore.  Furthermore, as we now use
fstream, doing the fopen to fopen64 remapping is now preventing the
fstream c++ headers to compile on el6 systems.

This patch just simply does away with the fopen to fopen64 remapping and
thus fixes the compilation on el6 systems.

	* src/abg-tools-utils.cc: Do not remap fopen to fopen64 as we
	don't use fopen explicitly anymore.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>