Correct various inconsequential typos.
In the leaf reporter member subtype changes are labelled as plain
changes and vice versa. This was probably due to the different
ordering of the code sections in the default and leaf reporters.
Output is unchanged as these tags currently map to the same strings.

When generating diff reports there were some rare cases where a
pretty representation might have been emitted twice or the trailing
whitespace might have been missing.

	* src/abg-leaf-reporter.cc (report): In the class_or_union_diff
	overoad, swap calls to report_mem_header to match the rest of the
	code.
	* src/abg-reporter-priv.cc (represent): In the var_diff_sptr
	overload, add some missing whitespace; remember we've emitted the
	pretty representation in 2 cases where this was omitted (though 1
	of these is the last case where it makes no difference).
	maybe_report_diff_for_symbol Add some missing whitespace; remember
	we've reported a diff (and need a trailing newline) in 1 case
	where this was omitted, also affecting the return value of the
	function (but no caller cares).

Signed-off-by: Giuliano Procida <gprocida@google.com>
Signed-off-by: Dodji Seketeli <dodji@redhat.com>